+++ integration_tests/tests/comprehendmedical.rs
@@ -1,0 +1,25 @@
+#![cfg(feature = "comprehendmedical")]
+
+extern crate rusoto_comprehendmedical;
+extern crate rusoto_core;
+
+use rusoto_comprehendmedical::{ComprehendMedical, ComprehendMedicalClient, DetectPHIRequest};
+use rusoto_core::Region;
+
+#[test]
+fn should_detect_sentiment() {
+    let client = ComprehendMedicalClient::new(Region::UsEast1);
+
+    let request = DetectPHIRequest {
+        text: "patient age is 42".to_owned(),
+        ..Default::default()
+    };
+
+    match client.detect_phi(request).sync() {
+        Ok(response) => {
+            println!("{:#?}", response);
+            assert_eq!(1, response.entities.len());
+        }
+        Err(err) => panic!("Expected OK response, got {:#?}", err),
+    };
+}